+++ renderer/+config.h.ts
@@ -1,0 +1,6 @@
+// See https://vike.dev/data-fetching
+export default {
+  clientRouting: true,
+  prefetchStaticAssets: 'viewport',
+  passToClient: ['pageProps', /* 'urlPathname', */ 'routeParams']
+}
