+++ renderer/+onHydrationEnd.ts
@@ -1,0 +1,5 @@
+function onHydrationEnd() {
+  // console.log('Hydration finished; page is now interactive.')
+}
+
+export { onHydrationEnd }
