+++ renderer/+onPageTransitionEnd.ts
@@ -1,0 +1,6 @@
+function onPageTransitionEnd() {
+  // console.log('Page transition end')
+  // document.body.classList.remove('page-transition')
+}
+
+export { onPageTransitionEnd }
