+++ renderer/+onPageTransitionStart.ts
@@ -1,0 +1,6 @@
+function onPageTransitionStart() {
+  // console.log('Page transition start')
+  // document.body.classList.add('page-transition')
+}
+  
+export { onPageTransitionStart }
